Move bignum implementation from bignum.rs to types.rs
